fix(README): remove tags hiding end of the content
I removed the `details` and `summary` HTML tags used to wrap
`./test/negative-tests/for-mdx/mdx_cases.ml.corrected.expected`'s
`part=matching` that were hiding the end of `README.mdx`.

PS: I don't know much about MDX, so I'm not sure if this is even a "fix"
or the best way to fix it. If you have a better solution, go for it!

Signed-off-by: GitFenixZ <soantonyly@gmail.com>